Identify the WPA metrics that have the biggest change between two periods.

This function uses the Information Value algorithm to predict which Workplace Analytics metrics are most explained by the change in dates.

IV_by_period(
  data,
  before_start = min(as.Date(data$Date, "%m/%d/%Y")),
  before_end,
  after_start = as.Date(before_end) + 1,
  after_end = max(as.Date(data$Date, "%m/%d/%Y")),
  mybins = 10,
  return = "table"
)

Arguments

data

Person Query as a dataframe including date column named "Date" This function assumes the data format is MM/DD/YYYY as is standard in a Workplace Analytics query output.

before_start

Start date of "before" time period in YYYY-MM-DD. Defaults to earliest date in dataset.

before_end

End date of "before" time period in YYYY-MM-DD

after_start

Start date of "after" time period in YYYY-MM-DD. Defaults to day after before_end.

after_end

End date of "after" time period in YYYY-MM-DD. Defaults to latest date in dataset.

mybins

Number of bins to cut the data into for Information Value analysis. Defaults to 10.

return

String specifying what to return. The current only valid option is "table".

Value

data frame containing all the variables and the corresponding Information Value.

See also

Author

Mark Powers mark.powers@microsoft.com

Examples

# \donttest{
# Returns a data frame
sq_data %>%
  IV_by_period(
    before_start = "2019-12-15",
    before_end = "2019-12-29",
    after_start = "2020-01-05",
    after_end = "2020-01-26"
  )
#>                                          Variable           IV
#> 1                                     Total_calls 5.153999e-01
#> 2                          Working_hours_in_calls 4.180817e-01
#> 3                                      Call_hours 4.168150e-01
#> 4              Generated_workload_calls_organized 1.602650e-01
#> 5            Generated_workload_call_participants 1.513890e-01
#> 6                   Generated_workload_call_hours 1.401184e-01
#> 7                            After_hours_in_calls 8.107796e-02
#> 8                           Internal_network_size 6.041966e-02
#> 9                                   Workweek_span 3.403181e-02
#> 10                 Working_hours_instant_messages 3.150579e-02
#> 11                          Instant_Message_hours 3.016572e-02
#> 12                      Working_hours_email_hours 2.743509e-02
#> 13      Generated_workload_instant_messages_hours 2.589020e-02
#> 14                   After_hours_instant_messages 2.456033e-02
#> 15                          Instant_messages_sent 2.343575e-02
#> 16                                       Meetings 2.119126e-02
#> 17                                    Email_hours 2.097932e-02
#> 18                                    Emails_sent 2.044890e-02
#> 19                        After_hours_email_hours 1.776537e-02
#> 20            Generated_workload_email_recipients 1.739243e-02
#> 21                 Generated_workload_email_hours 1.676688e-02
#> 22                          External_network_size 1.648850e-02
#> 23           Generated_workload_meeting_attendees 1.534064e-02
#> 24                      Low_quality_meeting_hours 1.513613e-02
#> 25                             Open_2_hour_blocks 1.338513e-02
#> 26                      After_hours_meeting_hours 1.278258e-02
#> 27       Redundant_meeting_hours__organizational_ 1.188721e-02
#> 28                            Collaboration_hours 1.164852e-02
#> 29                              Total_focus_hours 1.162128e-02
#> 30                   Collaboration_hours_external 1.157492e-02
#> 31               Generated_workload_meeting_hours 1.155354e-02
#> 32          Generated_workload_meetings_organized 1.101759e-02
#> 33                                  Meeting_hours 8.661404e-03
#> 34 Generated_workload_instant_messages_recipients 8.210957e-03
#> 35                After_hours_collaboration_hours 8.126027e-03
#> 36              Working_hours_collaboration_hours 7.423696e-03
#> 37                      Conflicting_meeting_hours 7.414662e-03
#> 38               Total_emails_sent_during_meeting 6.855551e-03
#> 39                     Networking_outside_company 6.738049e-03
#> 40                          Meetings_with_manager 6.268620e-03
#> 41             Meeting_hours_during_working_hours 6.143345e-03
#> 42                Time_in_self_organized_meetings 6.090035e-03
#> 43                              Open_1_hour_block 5.269519e-03
#> 44                     Multitasking_meeting_hours 4.897633e-03
#> 45                     Meeting_hours_with_manager 2.376759e-03
#> 46              Meeting_hours_with_manager_1_on_1 1.822441e-03
#> 47                   Meetings_with_manager_1_on_1 1.102444e-03
#> 48                  Manager_coaching_hours_1_on_1 2.192592e-04
#> 49                Networking_outside_organization 7.281065e-06
#> 50                       Meetings_with_skip_level 0.000000e+00
#> 51                  Meeting_hours_with_skip_level 0.000000e+00
#> 52          Redundant_meeting_hours__lower_level_ 0.000000e+00
#> 53                                          Layer 0.000000e+00
#> 54                                     HourlyRate 0.000000e+00
# }