Distribution of After-hours Collaboration Hours (Fizzy Drink plot)

Analyze weekly after-hours collaboration hours distribution, and returns a 'fizzy' scatter plot by default. Additional options available to return a table with distribution elements.

afterhours_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")

Arguments

data
A Standard Person Query dataset in the form of a data frame.

hrvar
HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"

mingroup
Numeric value setting the privacy threshold / minimum group size, defaults to 5.

return
Character vector specifying what to return, defaults to "plot". Valid inputs are "plot" and "table".

Details

Uses the metric After_hours_collaboration_hours. See create_fizz() for applying the same analysis to a different metric.

See also

Examples

afterhours_fizz(sq_data, hrvar = "Organization", return = "table")

#> # A tibble: 15 x 8
#>    group                 mean median    sd   min   max range     n
#>    <chr>                <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <int>
#>  1 Biz Dev               2.00   2.01 0.382  1.20  3.04  1.85    75
#>  2 Customer Service      2.78   2.37 1.38   1.48  7.38  5.90    61
#>  3 Facilities            2.85   2.40 1.06   1.41  5.10  3.68    72
#>  4 Finance-Corporate     2.36   1.97 1.02   1.21  4.85  3.64    68
#>  5 Finance-East          3.72   4.20 1.63   1.09  6.33  5.24    70
#>  6 Finance-South         2.94   2.33 1.35   1.29  5.85  4.56    81
#>  7 Finance-West          3.31   2.93 1.43   1.46  6.75  5.29    73
#>  8 Financial Planning    2.53   2.11 1.26   1.44  6.33  4.90    75
#>  9 G&A Central           2.93   2.44 1.19   1.19  5.07  3.88    57
#> 10 G&A East              2.24   2.17 0.540  1.31  3.91  2.59    65
#> 11 G&A South             3.05   2.95 1.08   1.24  5.52  4.28    76
#> 12 Human Resources       3.39   3.27 1.05   1.43  5.72  4.29    71
#> 13 Inventory Management  3.90   3.40 1.66   1.74  8.91  7.17    60
#> 14 IT-Corporate          3.33   2.66 1.75   1.15  7.24  6.08    68
#> 15 IT-East               2.63   2.49 0.904  1.30  5.47  4.16    62