After-hours Collaboration Time Trend - Line Chart

Provides a week by week view of after-hours collaboration time, visualized as line charts. By default returns a line chart for after-hours collaboration hours, with a separate panel per value in the HR attribute. Additional options available to return a summary table.

afterhours_line(data, hrvar = "Organization", mingroup = 5, return = "plot")

Arguments

data
A Standard Person Query dataset in the form of a data frame.

hrvar
HR Variable by which to split metrics, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"

mingroup
Numeric value setting the privacy threshold / minimum group size. Defaults to 5.

return
String specifying what to return. This must be one of the following strings:

"plot"

"table"

See Value for more information.

Value

A different output is returned depending on the value passed to the return argument:

"plot": ggplot object. A faceted line plot for the metric.

"table": data frame. A summary table for the metric.

Details

Uses the metric After_hours_collaboration_hours.

See also

Examples

# Return a line plot
afterhours_line(sq_data, hrvar = "LevelDesignation")

# Return summary table
afterhours_line(sq_data, hrvar = "LevelDesignation", return = "table")