Check a query to ensure that it is suitable for analysis

Prints diagnostic data about the data query to the R console, with information such as date range, number of employees, HR attributes identified, etc.

check_query(data, return = "message", validation = FALSE)

Arguments

data
A person-level query in the form of a data frame, including the standard person query, Ways of Working assessment query, and the hourly collaboration query.

return
Character string to specify whether to return a console message ("message"), a character string ("text"), or a data frame. Defaults to "message".

validation
Logical value to specify whether to return a check used by the validation_report(). Defaults to FALSE.

Details

This can be used with any person-level query, such as the standard person query, Ways of Working assessment query, and the hourly collaboration query. When run, this prints diagnostic data to the R console.

See also

Examples

check_query(sq_data)

#> 
#> The data used is a Person Query.
#> 
#> There are 1034 employees in this dataset.
#> 
#> Date ranges from 2019-11-03 to 2020-01-26.
#> 
#> There are 9 (estimated) HR attributes in the data:
#> `Domain`, `FunctionType`, `LevelDesignation`, `Region`, `Organization`, `attainment`, `TimeZone`, `IsInternal`, `IsActive`
#> 
#> There are 1034 active employees out of all in the dataset.
#> 
#> Variable name check:
#> 
#> `Collaboration_hours` is used instead of `Collaboration_hrs` in the data.
#> 
#> `Instant_Message_hours` is used instead of `Instant_message_hours` in the data.