Distribution of Collaboration Hours (Fizzy Drink plot)

Analyze weekly collaboration hours distribution, and returns a 'fizzy' scatter plot by default. Additional options available to return a table with distribution elements.

collaboration_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")

collab_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")

Arguments

data
A Standard Person Query dataset in the form of a data frame.

hrvar
HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"

mingroup
Numeric value setting the privacy threshold / minimum group size, defaults to 5.

return
Character vector specifying what to return, defaults to "plot". Valid inputs are "plot" and "table".

Details

Uses the metric Collaboration_hours.

See also

Examples

collaboration_fizz(sq_data, hrvar = "Organization", return = "table")

#> # A tibble: 15 x 8
#>    group                 mean median    sd   min   max range     n
#>    <chr>                <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <int>
#>  1 Biz Dev               14.7   15.0  2.32  8.46  19.9  11.4    75
#>  2 Customer Service      18.3   16.9  5.61 10.4   35.6  25.2    61
#>  3 Facilities            18.6   17.3  4.51 10.9   27.1  16.2    72
#>  4 Finance-Corporate     16.3   15.2  4.42  9.10  27.6  18.5    68
#>  5 Finance-East          22.2   25.1  7.33  9.44  33.5  24.1    70
#>  6 Finance-South         18.6   15.8  6.63  9.34  32.8  23.5    81
#>  7 Finance-West          20.6   19.6  6.00 11.7   33.5  21.9    73
#>  8 Financial Planning    17.1   15.9  5.09  9.49  34.2  24.7    75
#>  9 G&A Central           18.9   16.5  5.66 10.1   30.5  20.4    57
#> 10 G&A East              16.2   16.0  2.84 11.0   23.7  12.7    65
#> 11 G&A South             23.1   24.2  6.81 11.1   37.9  26.8    76
#> 12 Human Resources       24.4   25.0  6.20 10.8   35.8  25.0    71
#> 13 Inventory Management  26.2   25.0  6.18 12.1   41.7  29.6    60
#> 14 IT-Corporate          23.7   23.9  8.70  9.26  43.4  34.1    68
#> 15 IT-East               20.6   21.8  6.05  9.87  35.5  25.7    62