Generate a Connectivity report in HTML

The function generates an interactive HTML report using Standard Query data as an input. The report contains a series of summary analysis and visualisations relating to key connectivity metrics in Workplace Analytics, including external/internal network size vs breadth.

connectivity_report(
  data,
  hrvar = "LevelDesignation",
  mingroup = 5,
  path = "connectivity report",
  timestamp = TRUE
)

Arguments

data
A Standard Query dataset in the form of a data frame.

hrvar
HR Variable by which to split metrics, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"

mingroup
Numeric value setting the privacy threshold / minimum group size. Defaults to 5.

path
Pass the file path and the desired file name, excluding the file extension. For example, "collaboration report".

timestamp
Logical vector specifying whether to include a timestamp in the file name. Defaults to TRUE.

See also