Sample Hourly Collaboration data with Email and IMs

A sample dataset generated from an Hourly Collaboration query from the WpA demo tenant. The data is grouped by week and contains columns for both IMs sent and Emails sent.

em_data

Format

A data frame with 41567 rows and 56 variables:

PersonId

Date

IMs_sent_23_24

IMs_sent_22_23

IMs_sent_21_22

IMs_sent_20_21

IMs_sent_19_20

IMs_sent_18_19

IMs_sent_17_18

IMs_sent_16_17

IMs_sent_15_16

IMs_sent_14_15

IMs_sent_13_14

IMs_sent_12_13

IMs_sent_11_12

IMs_sent_10_11

IMs_sent_09_10

IMs_sent_08_09

IMs_sent_07_08

IMs_sent_06_07

IMs_sent_05_06

IMs_sent_04_05

IMs_sent_03_04

IMs_sent_02_03

IMs_sent_01_02

IMs_sent_00_01

Emails_sent_23_24

Emails_sent_22_23

Emails_sent_21_22

Emails_sent_20_21

Emails_sent_19_20

Emails_sent_18_19

Emails_sent_17_18

Emails_sent_16_17

Emails_sent_15_16

Emails_sent_14_15

Emails_sent_13_14

Emails_sent_12_13

Emails_sent_11_12

Emails_sent_10_11

Emails_sent_09_10

Emails_sent_08_09

Emails_sent_07_08

Emails_sent_06_07

Emails_sent_05_06

Emails_sent_04_05

Emails_sent_03_04

Emails_sent_02_03

Emails_sent_01_02

Emails_sent_00_01

Domain

LevelDesignation

Layer

Organization

IsInternal

IsActive

...

Source

https://workplaceanalytics-demo.office.com/en-us/Home