Extract date period

Return a data frame with the start and end date of the query data by default. There are options to return a descriptive string, which is used in the caption of plots in this package.

extract_date_range(data, return = "table")

Arguments

data
Data frame containing a query to pass through. The data frame must contain a Date column. Accepts a Person query or a Meeting query.

return
String specifying what output to return. Returns a table by default ("table"), but allows returning a descriptive string ("text").

Value

A different output is returned depending on the value passed to the return argument:

"table": data frame. A summary table containing the start and end date for the dataset.

"text": string. Contains a descriptive string on the start and end date for the dataset.

See also