Create count of distinct fields and percentage of employees with missing values for all HR variables

This function enables you to create a summary table to validate organizational data. This table will provide a summary of the data found in the Viva Insights Data sources page. This function will return a summary table with the count of distinct fields per HR attribute and the percentage of employees with missing values for that attribute. See hrvar_count() function for more detail on the specific HR attribute of interest.

hrvar_count_all(
  data,
  n_var = 50,
  return = "message",
  threshold = 100,
  maxna = 20,
  na_values = c("NA", "N/A", "#N/A", " ")
)

Arguments

data

A Standard Person Query dataset in the form of a data frame.

n_var

number of HR variables to include in report as rows. Default is set to 50 HR variables.

return

String to specify what to return

threshold

The max number of unique values allowed for any attribute. Default is 100.

maxna

The max percentage of NAs allowable for any column. Default is 20.

na_values

Character vector of values to be treated as missing. Default is c("NA", "N/A", "#N/A", " ").

Value

Returns an error message by default, where 'text' is passed in return.

'table': data frame. A summary table listing the number of distinct fields and percentage of missing values for the specified number of HR attributes will be returned.

'message': outputs a message indicating which values are beyond the specified thresholds.

Note

As of v1.6.3, the function can detect and report text values like "NA", "N/A", "#N/A", and spaces that represent missing values, by treating them as NA values. You can customize which values are treated as missing with the na_values parameter. This can be validated as per:

dv_data %>%
  mutate(TempOrg = sample(c("NA", "#N/A", " "), size = nrow(.), replace = TRUE)) %>%
  hrvar_count_all(return = "table")

See also

Examples

# Return a summary table of all HR attributes
hrvar_count_all(sq_data, return = "table")
#> # A tibble: 9 × 4
#>   Attributes       `Unique values` `Total missing values` `% missing values`
#>   <chr>                      <dbl>                  <dbl>              <dbl>
#> 1 Domain                         1                      0                  0
#> 2 FunctionType                   8                      0                  0
#> 3 IsActive                       1                      0                  0
#> 4 IsInternal                     1                      0                  0
#> 5 LevelDesignation               6                      0                  0
#> 6 Organization                   5                      0                  0
#> 7 Region                         1                      0                  0
#> 8 TimeZone                       1                      0                  0
#> 9 attainment                     0                   4403                100