Tenure calculation based on different input dates, returns data summary table or histogram

This function calculates employee tenure based on different input dates. identify_tenure uses the latest Date available if user selects "Date", but also have flexibility to select a specific date, e.g. "1/1/2020".

identify_tenure(
  data,
  end_date = "Date",
  beg_date = "HireDate",
  maxten = 40,
  return = "message"
)

Arguments

data
A Standard Query dataset in the form of a data frame.

end_date
A string specifying the name of the date variable representing the latest date. Defaults to "Date".

beg_date
A string specifying the name of the date variable representing the hire date. Defaults to "HireDate".

maxten
A numeric value representing the maximum tenure. If the tenure exceeds this threshold, it would be accounted for in the flag message.

return
String to specify what to return. Defaults to "message". Other valid values include "text", "plot", "data_cleaned", "data_dirty", and "data". For "data", a data frame with the PersonId and a calculated variable called TenureYear is returned.

See also

Examples

if (FALSE) {
# Add HireDate to sq_data - method #1
sq_data2 <- sq_data %>% mutate(HireDate = as.Date("1/1/2015", format = "%m/%d/%Y" ))
identify_tenure(sq_data2)

# Add HireDate to sq_data - method #2
sq_data$HireDate <-
  rep(sample(seq(as.Date('1975/01/01'),
                 as.Date('2019/11/01'), by="day"), 15119),2)

identify_tenure(sq_data)
}