Plot a Sample of Working Patterns using Flexibility Index output

This is a helper function for plotting visualizations for the Flexibility Index using the data output from flex_index(). This is used within flex_index() itself as an internal function.

plot_flex_index(
  data,
  sig_label = "Signals_sent_",
  method = "sample",
  start_hour = 9,
  end_hour = 17
)

Arguments

data
Data frame. Direct data output from flex_index().

sig_label
Character string for identifying signal labels.

method
Character string for determining which plot to return. Options include "sample", "common", and "time". "sample" plots a sample of ten working patterns; "common" plots the ten most common working patterns; "time" plots the Flexibility Index for the group over time.

start_hour
See flex_index().

end_hour
See flex_index().

See also

Examples

if (FALSE) {
# Examples of how to test the plotting options individually
# Sample of 10 work patterns
em_data %>%
  flex_index(return = "data") %>%
  plot_flex_index(method = "sample")

# 10 most common work patterns
em_data %>%
  flex_index(return = "data") %>%
  plot_flex_index(method = "common")

# Plot Flexibility Index over time
em_data %>%
  flex_index(return = "data") %>%
  plot_flex_index(method = "time")

}