Internal function for plotting the hourly activity patterns.

This is used within plot_flex_index() and workpatterns_rank().

plot_hourly_pat(
  data,
  start_hour,
  end_hour,
  legend,
  legend_label,
  legend_text = "Observed activity",
  rows,
  title,
  subtitle,
  caption,
  ylab = paste("Top", rows, "activity patterns")
)

Arguments

data

Data frame containing three columns:

patternRank

Hours

Freq

start_hour

Numeric value to specify expected start hour.

end_hour

Numeric value to specify expected end hour.

legend

Data frame containing the columns:

patternRank

Any column to be used in the grey label box, supplied to legend_label

legend_label

String specifying column to display in the grey label box

legend_text

String to be used in the bottom legend label.

rows

Number of rows to show in plot.

title

String to specify plot title.

subtitle

String to specify plot subtitle.

caption

String to specify plot caption.

ylab

String to specify plot y-axis label.