Sample Standard Query dataset

A dataset generated from a Standard Query from WpA.

sq_data

Format

A data frame with 13442 rows and 66 variables:

PersonId

Date

Workweek_span

Time between the person's first sent email or meeting attended and the last email or meeting for each day of the work week.

Meetings_with_skip_level

Meeting_hours_with_skip_level

Generated_workload_email_hours

Generated_workload_email_recipients

Generated_workload_instant_messages_hours

Generated_workload_instant_messages_recipients

Generated_workload_call_hours

Generated_workload_call_participants

Generated_workload_calls_organized

External_network_size

Internal_network_size

Networking_outside_company

Networking_outside_organization

After_hours_meeting_hours

Open_1_hour_block

Open_2_hour_blocks

Total_focus_hours

Low_quality_meeting_hours

Total_emails_sent_during_meeting

Meetings

Meeting_hours

Conflicting_meeting_hours

Multitasking_meeting_hours

Redundant_meeting_hours__lower_level_

Redundant_meeting_hours__organizational_

Time_in_self_organized_meetings

Meeting_hours_during_working_hours

Generated_workload_meeting_attendees

Generated_workload_meeting_hours

Generated_workload_meetings_organized

Manager_coaching_hours_1_on_1

Meetings_with_manager

Meeting_hours_with_manager

Meetings_with_manager_1_on_1

Meeting_hours_with_manager_1_on_1

After_hours_email_hours

Emails_sent

Email_hours

Number of hours the person spent sending and receiving emails.

Working_hours_email_hours

After_hours_instant_messages

Instant_messages_sent

Instant_Message_hours

Working_hours_instant_messages

After_hours_collaboration_hours

Collaboration_hours

Collaboration_hours_external

Working_hours_collaboration_hours

After_hours_in_calls

Total_calls

Call_hours

Working_hours_in_calls

Domain

FunctionType

LevelDesignation

Layer

Region

Organization

zId

attainment

TimeZone

HourlyRate

IsInternal

IsActive

...

Source

https://workplaceanalytics-demo.office.com/en-us/Home