Standardise variable names to a Standard Person Query

This function standardises the variable names to a Standard Person Query, where the standard use case is to pass a Ways of Working Assessment Query to the function.

standardise_pq(data)

standardize_pq(data)

Arguments

data

A Ways of Working Assessment query to pass through as a data frame.

Value

data frame containing the formatted query passed to the function.

Details

The following standardisation steps are taken:

Collaboration_hrs -> Collaboration_hours

Instant_message_hours -> Instant_Message_hours

See also