Distribution of Work Week Span (Fizzy Drink plot)

Analyze Work Week Span distribution, and returns a 'fizzy' scatter plot by default. Additional options available to return a table with distribution elements.

workloads_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")

Arguments

data
A Standard Person Query dataset in the form of a data frame.

hrvar
HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"

mingroup
Numeric value setting the privacy threshold / minimum group size, defaults to 5.

return
Character vector specifying what to return, defaults to "plot". Valid inputs are "plot" and "table".

See also

Examples

meeting_fizz(sq_data, hrvar = "Organization", return = "table")

#> # A tibble: 15 x 8
#>    group                 mean median    sd   min   max range     n
#>    <chr>                <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <int>
#>  1 Biz Dev               7.02   6.96  1.49  2.98  10.5  7.51    75
#>  2 Customer Service      8.78   7.66  3.50  5     20.4 15.4     61
#>  3 Facilities            9.39   9.16  2.30  4.41  13.9  9.49    72
#>  4 Finance-Corporate     8.16   7.43  2.50  3.83  15.7 11.9     68
#>  5 Finance-East         11.4   12.3   3.98  4.61  18.2 13.6     70
#>  6 Finance-South         9.27   8.03  3.68  3.23  16.9 13.7     81
#>  7 Finance-West         10.4    9.98  3.23  4.35  19.5 15.2     73
#>  8 Financial Planning    8.36   7.99  2.92  3.25  17.6 14.3     75
#>  9 G&A Central           9.28   8.02  3.07  4.48  16.2 11.7     57
#> 10 G&A East              7.84   7.5   1.64  4.67  12.0  7.33    65
#> 11 G&A South            11.6   12.0   3.96  5.31  20.9 15.6     76
#> 12 Human Resources      12.6   12.6   3.87  5.10  20.0 14.9     71
#> 13 Inventory Management 13.7   12.9   3.46  6.26  22.0 15.7     60
#> 14 IT-Corporate         12.3   12.5   5.24  3.65  24.9 21.2     68
#> 15 IT-East              10.8   11.4   3.72  4.21  19.9 15.7     62