Identify the WPA metrics that have the biggest change between two periods.

This function uses the Information Value algorithm to predict which Workplace Analytics metrics are most explained by the change in dates.

IV_by_period(
  data,
  before_start = min(as.Date(data$Date, "%m/%d/%Y")),
  before_end,
  after_start = as.Date(before_end) + 1,
  after_end = max(as.Date(data$Date, "%m/%d/%Y")),
  mybins = 10,
  return = "table"
)

Arguments

data
Person Query as a dataframe including date column named "Date" This function assumes the data format is MM/DD/YYYY as is standard in a Workplace Analytics query output.

before_start
Start date of "before" time period in YYYY-MM-DD. Defaults to earliest date in dataset.

before_end
End date of "before" time period in YYYY-MM-DD

after_start
Start date of "after" time period in YYYY-MM-DD. Defaults to day after before_end.

after_end
End date of "after" time period in YYYY-MM-DD. Defaults to latest date in dataset.

mybins
Number of bins to cut the data into for Information Value analysis. Defaults to 10.

return
String specifying what to return. The current only valid option is "table".

Value

data frame containing all the variables and the corresponding Information Value.

See also

Author

Mark Powers mark.powers@microsoft.com

Examples

# Returns a data frame
sq_data %>%
  IV_by_period(
    before_start = "2019-11-03",
    before_end = "2019-11-09",
    after_start = "2019-11-10",
    after_end = "2019-11-16"
  )

#>                                          Variable           IV
#> 1                       After_hours_meeting_hours 4.063198e-02
#> 2                 After_hours_collaboration_hours 3.214857e-02
#> 3                         After_hours_email_hours 2.827746e-02
#> 4                    After_hours_instant_messages 2.646030e-02
#> 5                           Instant_messages_sent 2.390204e-02
#> 6             Generated_workload_email_recipients 2.351152e-02
#> 7                Total_emails_sent_during_meeting 2.253198e-02
#> 8                           Meetings_with_manager 2.037033e-02
#> 9                               Open_1_hour_block 2.031148e-02
#> 10                                  Workweek_span 1.875483e-02
#> 11                                    Email_hours 1.847501e-02
#> 12                      Low_quality_meeting_hours 1.840299e-02
#> 13                          Instant_Message_hours 1.832036e-02
#> 14      Generated_workload_instant_messages_hours 1.807253e-02
#> 15                      Conflicting_meeting_hours 1.673358e-02
#> 16                                       Meetings 1.621851e-02
#> 17             Meeting_hours_during_working_hours 1.607176e-02
#> 18                     Meeting_hours_with_manager 1.580855e-02
#> 19                                  Meeting_hours 1.547852e-02
#> 20                     Multitasking_meeting_hours 1.512573e-02
#> 21                            Collaboration_hours 1.431442e-02
#> 22                Time_in_self_organized_meetings 1.366691e-02
#> 23                 Generated_workload_email_hours 1.362824e-02
#> 24                                     Call_hours 1.345850e-02
#> 25                         Working_hours_in_calls 1.338040e-02
#> 26                      Working_hours_email_hours 1.317128e-02
#> 27 Generated_workload_instant_messages_recipients 1.222653e-02
#> 28                 Working_hours_instant_messages 1.194240e-02
#> 29                           After_hours_in_calls 1.160868e-02
#> 30                   Collaboration_hours_external 1.105485e-02
#> 31       Redundant_meeting_hours__organizational_ 1.065584e-02
#> 32                              Total_focus_hours 1.064046e-02
#> 33                                    Emails_sent 1.023741e-02
#> 34               Generated_workload_meeting_hours 1.021543e-02
#> 35                          External_network_size 9.104338e-03
#> 36                             Open_2_hour_blocks 8.839182e-03
#> 37                          Internal_network_size 8.784474e-03
#> 38              Working_hours_collaboration_hours 8.124579e-03
#> 39           Generated_workload_call_participants 7.956155e-03
#> 40                  Generated_workload_call_hours 6.901190e-03
#> 41             Generated_workload_calls_organized 6.404576e-03
#> 42                                    Total_calls 4.848446e-03
#> 43                   Meetings_with_manager_1_on_1 4.710288e-03
#> 44          Generated_workload_meetings_organized 4.059213e-03
#> 45           Generated_workload_meeting_attendees 3.425110e-03
#> 46              Meeting_hours_with_manager_1_on_1 3.214828e-03
#> 47                     Networking_outside_company 1.235318e-03
#> 48                  Manager_coaching_hours_1_on_1 2.339200e-04
#> 49                Networking_outside_organization 4.851778e-06
#> 50                       Meetings_with_skip_level 0.000000e+00
#> 51                  Meeting_hours_with_skip_level 0.000000e+00
#> 52          Redundant_meeting_hours__lower_level_ 0.000000e+00
#> 53                                          Layer 0.000000e+00
#> 54                                     HourlyRate 0.000000e+00