Create a sankey chart from a two-column count table

Create a networkD3 style sankey chart based on a long count table with two variables. The input data should have three columns, where each row is a unique group:

Variable 1

Variable 2

Count

create_sankey(data, var1, var2, count = "n")

Arguments

data
Data frame of the long count table.

var1
String containing the name of the variable to be shown on the left.

var2
String containing the name of the variable to be shown on the right.

count
String containing the name of the count variable.

Examples

# \donttest{
sq_data %>%
  dplyr::count(Organization, FunctionType) %>%
  create_sankey(var1 = "Organization", var2 = "FunctionType")
# }