Identify employees who have churned from the dataset

This function identifies and counts the number of employees who have churned from the dataset by measuring whether an employee who is present in the first n (n1) weeks of the data is present in the last n (n2) weeks of the data.

identify_churn(data, n1 = 6, n2 = 6, return = "message", flip = FALSE)

Arguments

data
A Person Query as a data frame. Must contain a PersonId.

n1
A numeric value specifying the number of weeks at the beginning of the period that defines the measured employee set. Defaults to 6.

n2
A numeric value specifying the number of weeks at the end of the period to calculate whether employees have churned from the data. Defaults to 6.

return
String specifying what to return. Defaults to "message", with options to return a character string ("text") or the PersonId of employees who have been identified as churned ("data").

flip
Logical, defaults to FALSE. This determines whether to reverse the logic of identifying the non-overlapping set. If set to TRUE, this effectively identifies new-joiners, or those who were not present in the first n weeks of the data but were present in the final n weeks.

Details

An additional use case of this function is the ability to identify "new-joiners" by using the argument flip.

If an employee is present in the first n weeks of the data but not present in the last n weeks of the data, the function considers the employee as churned. As the measurement period is defined by the number of weeks from the start and the end of the passed data frame, you may consider filtering the dates accordingly before running this function.

Another assumption that is in place is that any employee whose PersonId is not available in the data has churned. Note that there may be other reasons why an employee's PersonId may not be present, e.g. maternity/paternity leave, Workplace Analytics license has been removed, shift to a low-collaboration role (to the extent that he/she becomes inactive).

Examples

sq_data %>% identify_churn(n1 = 3, n2 = 3, return = "message")

#> Churn:
#> There are 0 employees from 2019-11-03 to 2019-11-17 (3 weeks) who are no longer present in 2020-01-12 to 2020-01-26 (3 weeks).