Extract top low-engagement meetings from the Meeting Query

Pass a Standard Meeting Query and extract the top low engagement meetings.

meeting_extract(
  data,
  recurring_only = TRUE,
  top_n = 30,
  fte_month = 180,
  fte_week = 40,
  return = "table"
)

Arguments

data
Data frame containing a Standard Meeting Query to pass through.

recurring_only
Logical value indicating whether to only filter by recurring meetings.

top_n
Numeric value for the top number of results to return in the output.

fte_month
Numeric value for the assumed number of employee hours per month for conversion calculations. Defaults to 180.

fte_week
Numeric value for the assumed number of employee hours per week for conversion calculations. Defaults to 180.

return
String specifying what to return. This must be one of the following strings:

"table"

"data"

See Value for more information.

Value

A different output is returned depending on the value passed to the return argument:

"table": data frame. A summary table containing the top n low engagement meetings

"data": data frame. Contains the full computed metrics related to the top n low engagement meetings

See also

Examples

meeting_extract(mt_data,
                recurring_only = FALSE,
                top_n = 10,
                return = "table")
#>                                                  Subject Attendee hours
#> 1                            ravi prep and dev interview           6.20
#> 2  visit new position with every interview in enterprise           6.00
#> 3               volometrix keynote from david enterprise           5.25
#> 4       staff product and karen strategic commissions up           5.00
#> 5        transition every parking on reporting cafeteria           4.50
#> 6                                       update all visit           3.50
#> 7                                             ryan apple           3.00
#> 8                               todd transfer project of           3.00
#> 9   updated messaging up the budget and every discussion           3.00
#> 10                            weekly status up reporting           3.00
#>    Average emails per attendee hour Duration Attendees Number of recurrences
#> 1                          3.064516     2.50         5                     1
#> 2                          2.000000     1.00         6                     1
#> 3                          2.476190     0.75         7                     1
#> 4                          2.200000     1.25         6                     1
#> 5                          2.222222     0.50        10                     1
#> 6                          2.000000     0.50         7                     1
#> 7                          2.000000     1.00         3                     1
#> 8                          2.000000     1.00         6                     1
#> 9                          2.000000     1.00         4                     1
#> 10                         2.666667     0.75         4                     1
#>    Average Attendance Rate Invitees Equivalent FTE month Equivalent FTE week
#> 1                0.8333333        6           0.03444444             0.15500
#> 2                0.8571429        7           0.03333333             0.15000
#> 3                1.0000000        7           0.02916667             0.13125
#> 4                0.8571429        7           0.02777778             0.12500
#> 5                0.9090909       11           0.02500000             0.11250
#> 6                1.0000000        7           0.01944444             0.08750
#> 7                1.0000000        3           0.01666667             0.07500
#> 8                0.8571429        7           0.01666667             0.07500
#> 9                1.0000000        4           0.01666667             0.07500
#> 10               1.0000000        4           0.01666667             0.07500