Generate a Data Validation report in HTML

The function generates an interactive HTML report using Standard Person Query data as an input. The report contains checks on Workplace Analytics query outputs to provide diagnostic information for the Analyst prior to analysis.

An additional Standard Meeting Query can be provided to perform meeting subject line related checks. This is optional and the validation report can be run without it.

validation_report(
  data,
  meeting_data = NULL,
  hrvar = "Organization",
  path = "validation report",
  hrvar_threshold = 150,
  timestamp = TRUE
)

Arguments

data
A Standard Person Query dataset in the form of a data frame.

meeting_data
An optional Meeting Query dataset in the form of a data frame.

hrvar
HR Variable by which to split metrics, defaults to "Organization" but accepts any character vector, e.g. "Organization"

path
Pass the file path and the desired file name, excluding the file extension.

hrvar_threshold
Numeric value determining the maximum number of unique values to be allowed to qualify as a HR variable. This is passed directly to the threshold argument within hrvar_count_all().

timestamp
Logical vector specifying whether to include a timestamp in the file name. Defaults to TRUE.

Value

An HTML report with the same file name as specified in the arguments is generated in the working directory. No outputs are directly returned by the function.

Details

For your input to data or meeting_data, please use the function wpa::import_wpa() to import your csv query files into R. This function will standardize format and prepare the data as input for this report.

If you are passing a Ways of Working Assessment query instead of a Standard Person query to the data argument, please also use standardise_pq() to make the variable names consistent with a Standard Person Query.

Checking functions within validation_report()

check_query()

flag_ch_ratio()

hrvar_count_all()

identify_privacythreshold()

identify_nkw()

identify_holidayweeks()

subject_validate()

identify_tenure()

flag_outlooktime()

identify_shifts()

track_HR_change()

You can browse each individual function for details on calculations.

See also

Examples

# \donttest{
validation_report(dv_data,
                  meeting_data = mt_data,
                  hrvar = "Organization")

#> 
#> 
#> processing file: validation-report-20210318_172029.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |.                                                                     |   1%
#>   ordinary text without R code
#> 
#>   |                                                                              |.                                                                     |   2%
#> label: unnamed-chunk-1 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |..                                                                    |   3%
#>   ordinary text without R code
#> 
#>   |                                                                              |...                                                                   |   4%
#> label: unnamed-chunk-2 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |...                                                                   |   5%
#>   ordinary text without R code
#> 
#>   |                                                                              |....                                                                  |   6%
#> label: unnamed-chunk-3 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#> 
#>   ordinary text without R code
#> 
#>   |                                                                              |.....                                                                 |   7%
#> label: unnamed-chunk-4 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |......                                                                |   8%
#>   ordinary text without R code
#> 
#>   |                                                                              |......                                                                |   9%
#> label: unnamed-chunk-5 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 

#>   |                                                                              |.......                                                               |  10%
#>   ordinary text without R code
#> 
#>   |                                                                              |........                                                              |  11%
#> label: unnamed-chunk-6 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |........                                                              |  12%
#>   ordinary text without R code
#> 
#>   |                                                                              |.........                                                             |  13%
#> label: unnamed-chunk-7 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |..........                                                            |  14%
#>   ordinary text without R code
#> 
#>   |                                                                              |..........                                                            |  15%
#> label: unnamed-chunk-8 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |...........                                                           |  16%
#>   ordinary text without R code
#> 
#>   |                                                                              |............                                                          |  17%
#> label: unnamed-chunk-9 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#> 
#>   ordinary text without R code
#> 
#>   |                                                                              |.............                                                         |  18%
#> label: unnamed-chunk-10 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.............                                                         |  19%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............                                                        |  20%
#> label: unnamed-chunk-11 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |...............                                                       |  21%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............                                                       |  22%
#> label: unnamed-chunk-12 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |................                                                      |  23%
#>   ordinary text without R code
#> 
#>   |                                                                              |.................                                                     |  24%
#> label: unnamed-chunk-13 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.................                                                     |  25%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................                                                    |  26%
#> label: unnamed-chunk-14 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |...................                                                   |  27%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................                                                   |  28%
#> label: unnamed-chunk-15 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |....................                                                  |  28%
#>   ordinary text without R code
#> 
#>   |                                                                              |.....................                                                 |  29%
#> label: unnamed-chunk-16 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.....................                                                 |  30%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................                                                |  31%
#> label: unnamed-chunk-17 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |......................                                                |  32%
#>   ordinary text without R code
#> 
#>   |                                                                              |.......................                                               |  33%
#> label: unnamed-chunk-18 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |........................                                              |  34%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................                                              |  35%
#> label: unnamed-chunk-19 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.........................                                             |  36%
#>   ordinary text without R code
#> 
#>   |                                                                              |..........................                                            |  37%
#> label: unnamed-chunk-20 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |..........................                                            |  38%
#>   ordinary text without R code
#> 
#>   |                                                                              |...........................                                           |  39%
#> label: unnamed-chunk-21 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |............................                                          |  39%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: unnamed-chunk-22 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 

#>   |                                                                              |.............................                                         |  41%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............................                                        |  42%
#> label: unnamed-chunk-23 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |..............................                                        |  43%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................                                       |  44%
#> label: unnamed-chunk-24 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |...............................                                       |  45%
#>   ordinary text without R code
#> 
#>   |                                                                              |................................                                      |  46%
#> label: unnamed-chunk-25 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.................................                                     |  47%
#>   ordinary text without R code
#> 
#>   |                                                                              |.................................                                     |  48%
#> label: unnamed-chunk-26 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |..................................                                    |  49%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................................                                   |  50%
#> label: unnamed-chunk-27 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#> 
#>   ordinary text without R code
#> 
#>   |                                                                              |....................................                                  |  51%
#> label: unnamed-chunk-28 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.....................................                                 |  52%
#>   ordinary text without R code
#> 
#>   |                                                                              |.....................................                                 |  53%
#> label: unnamed-chunk-29 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |......................................                                |  54%
#>   ordinary text without R code
#> 
#>   |                                                                              |.......................................                               |  55%
#> label: unnamed-chunk-30 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 

#>   |                                                                              |.......................................                               |  56%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................                              |  57%
#> label: unnamed-chunk-31 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |........................................                              |  58%
#>   ordinary text without R code
#> 
#>   |                                                                              |.........................................                             |  59%
#> label: unnamed-chunk-32 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |..........................................                            |  60%
#>   ordinary text without R code
#> 
#>   |                                                                              |..........................................                            |  61%
#> label: unnamed-chunk-33 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 

#>   |                                                                              |...........................................                           |  61%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................................                          |  62%
#> label: unnamed-chunk-34 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |............................................                          |  63%
#>   ordinary text without R code
#> 
#>   |                                                                              |.............................................                         |  64%
#> label: unnamed-chunk-35 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |..............................................                        |  65%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............................................                        |  66%
#> label: unnamed-chunk-36 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |...............................................                       |  67%
#>   ordinary text without R code
#> 
#>   |                                                                              |................................................                      |  68%
#> label: unnamed-chunk-37 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |................................................                      |  69%
#>   ordinary text without R code
#> 
#>   |                                                                              |.................................................                     |  70%
#> label: unnamed-chunk-38 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.................................................                     |  71%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................................................                    |  72%
#> label: unnamed-chunk-39 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |...................................................                   |  72%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................................................                   |  73%
#> label: unnamed-chunk-40 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 

#>   |                                                                              |....................................................                  |  74%
#>   ordinary text without R code
#> 
#>   |                                                                              |.....................................................                 |  75%
#> label: unnamed-chunk-41 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.....................................................                 |  76%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................................................                |  77%
#> label: unnamed-chunk-42 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.......................................................               |  78%
#>   ordinary text without R code
#> 
#>   |                                                                              |.......................................................               |  79%
#> label: unnamed-chunk-43 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |........................................................              |  80%
#>   ordinary text without R code
#> 
#>   |                                                                              |.........................................................             |  81%
#> label: unnamed-chunk-44 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.........................................................             |  82%
#>   ordinary text without R code
#> 
#>   |                                                                              |..........................................................            |  83%
#> label: unnamed-chunk-45 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#> 
#>   ordinary text without R code
#> 
#>   |                                                                              |...........................................................           |  84%
#> label: unnamed-chunk-46 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |............................................................          |  85%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................................................          |  86%
#> label: unnamed-chunk-47 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.............................................................         |  87%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............................................................        |  88%
#> label: unnamed-chunk-48 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |..............................................................        |  89%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................................       |  90%
#> label: unnamed-chunk-49 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |................................................................      |  91%
#>   ordinary text without R code
#> 
#>   |                                                                              |................................................................      |  92%
#> label: unnamed-chunk-50 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |.................................................................     |  93%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................................................................    |  94%
#> label: unnamed-chunk-51 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#> 
#>   ordinary text without R code
#> 
#>   |                                                                              |...................................................................   |  95%
#> label: unnamed-chunk-52 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |...................................................................   |  96%
#>   ordinary text without R code
#> 
#>   |                                                                              |....................................................................  |  97%
#> label: unnamed-chunk-53 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |..................................................................... |  98%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................................................................... |  99%
#> label: unnamed-chunk-54 (with options) 
#> List of 3
#>  $ echo      : logi FALSE
#>  $ fig.height: num 9
#>  $ fig.width : num 12
#> 
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: validation-report-20210318_172029.knit.md

#> /usr/local/bin/pandoc +RTS -K512m -RTS validation-report-20210318_172029.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output validation-report-20210318_172029.html --lua-filter /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --self-contained --variable bs3=TRUE --standalone --section-divs --table-of-contents --toc-depth 3 --variable toc_float=1 --variable toc_selectors=h1,h2,h3 --variable toc_smooth_scroll=1 --variable toc_print=1 --template /Users/runner/work/_temp/Library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --variable theme=cosmo --include-in-header /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpNCn7m5/rmarkdown-str74a180868f2.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' 

#> 
#> Output created: validation-report-20210318_172029.html

# }