Generate a Wellbeing Report in HTML

Generate a static HTML report on wellbeing, taking a custom Wellbeing Query and an Hourly Collaboration query as inputs. See Required metrics section for more details on the required inputs for the Wellbeing Query. Note that this function is currently still in experimental/development stage and may experience changes in the near term.

wellbeing_report(
  wbq,
  hcq,
  hrvar = "Organization",
  mingroup = 5,
  start_hour = "0900",
  end_hour = "1700",
  path = "wellbeing_report"
)

Arguments

wbq

Data frame. A custom Wellbeing Query dataset based on the Person Query. If certain metrics are missing from the Wellbeing / Person Query, the relevant visual will show up with an indicative message.

hcq

Data frame. An Hourly Collaboration Query dataset.

hrvar

String specifying HR attribute to cut by archetypes. Defaults to Organization.

mingroup

Numeric value setting the privacy threshold / minimum group size. Defaults to 5.

start_hour

A character vector specifying starting hours, e.g. "0900". Note that this currently only supports hourly increments. If the official hours specifying checking in and 9 AM and checking out at 5 PM, then "0900" should be supplied here.

end_hour

A character vector specifying starting hours, e.g. "1700". Note that this currently only supports hourly increments. If the official hours specifying checking in and 9 AM and checking out at 5 PM, then "1700" should be supplied here.

path

Pass the file path and the desired file name, excluding the file extension. Defaults to "wellbeing_report".

Required metrics

A full list of the required metrics are as follows:

Urgent_meeting_hours

IMs_sent_other_level

IMs_sent_same_level

Emails_sent_other_level

Emails_sent_same_level

Emails_sent

IMs_sent

Meeting_hours_intimate_group

Meeting_hours_1on1

Urgent_email_hours

Unscheduled_call_hours

Meeting_hours

Instant_Message_hours

Email_hours

Total_focus_hours

Weekend_IMs_sent

Weekend_emails_sent

After_hours_collaboration_hours

After_hours_meeting_hours

After_hours_instant_messages

After_hours_in_unscheduled_calls

After_hours_email_hours

Collaboration_hours

Workweek_span